requirements.txt
streamlit
streamlit-lottie
requests
pdfminer.six
sentence-transformers
scikit-learn
numpy
pandas
